+++ adapter/rest-api/api-dto/src/main/java/com/di/hostme/service/rest/api/dto/country/CountryDetailResponse.java
@@ -1,0 +1,11 @@
+package com.di.hostme.service.rest.api.dto.country;
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.immutables.value.Value;
+
+@JsonSerialize(as = ImmutableCountryDetailResponse.class)
+@JsonDeserialize(as = ImmutableCountryDetailResponse.class)
+@Value.Immutable
+public sealed interface CountryDetailResponse extends BaseCountryResponse
+    permits ImmutableCountryDetailResponse, ImmutableCountryDetailResponse.Json {}