+++ adapter/rest-api/api-dto/src/main/java/com/di/hostme/service/rest/api/dto/package-info.java
@@ -1,0 +1,4 @@
+@Value.Style(stagedBuilder = true)
+package com.di.hostme.service.rest.api.dto;
+
+import org.immutables.value.Value;