+++ adapter/rest-api/src/main/java/com/di/hostme/service/error/ServiceAdviceController.java
@@ -1,9 +1,9 @@
 package com.di.hostme.service.error;
 
 import com.di.hostme.service.controller.BaseController;
-import com.di.hostme.service.domain.common.error.BusinessError;
-import com.di.hostme.service.domain.common.error.BusinessException;
-import com.di.hostme.service.rest.api.dto.common.error.ErrorResponseList;
+import com.di.hostme.service.domain.shared.common.error.BusinessError;
+import com.di.hostme.service.domain.shared.common.error.BusinessException;
+import com.di.hostme.service.rest.api.dto.common.error.ErrorResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
@@ -16,12 +16,12 @@ public class ServiceAdviceController {
   public static final Logger log = LoggerFactory.getLogger(ServiceAdviceController.class);
 
   @ExceptionHandler(BusinessException.class)
-  public ResponseEntity<ErrorResponseList> handle(BusinessException e) {
+  public ResponseEntity<ErrorResponse> handle(BusinessException e) {
     log.error("Handle '{}'.", e.getClass().getName(), e);
     return mapErrorToResponse(e.getBusinessError());
   }
 
-  public ResponseEntity<ErrorResponseList> mapErrorToResponse(final BusinessError businessError) {
+  public ResponseEntity<ErrorResponse> mapErrorToResponse(final BusinessError businessError) {
     return BaseController.mapErrorToResponse(businessError);
   }
 }