+++ adapter/rest-api/src/main/java/com/di/hostme/service/mapper/ErrorResponseMapper.java
@@ -1,31 +1,29 @@
 package com.di.hostme.service.mapper;
 
-import com.di.hostme.service.domain.common.error.BusinessError;
+import com.di.hostme.service.domain.shared.common.error.BusinessError;
 import com.di.hostme.service.rest.api.dto.common.error.ErrorResponse;
-import com.di.hostme.service.rest.api.dto.common.error.ErrorResponseList;
+import com.di.hostme.service.rest.api.dto.common.error.ErrorResponseFault;
 import com.di.hostme.service.rest.api.dto.common.error.ImmutableErrorResponse;
-import com.di.hostme.service.rest.api.dto.common.error.ImmutableErrorResponseList;
+import com.di.hostme.service.rest.api.dto.common.error.ImmutableErrorResponseFault;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 
-public class ErrorResponseMapper {
+public final class ErrorResponseMapper {
 
-  public static ResponseEntity<ErrorResponseList> mapErrorToResponse(
+  public static ResponseEntity<ErrorResponse> mapErrorToResponse(
       BusinessError businessError, String serviceName) {
-    final ErrorResponse errorResponse =
-        ImmutableErrorResponse.builder()
-            .faultCode(businessError.errorCode().toString())
-            .faultMessage(businessError.errorCode().template())
-            .addAllFaultMessageParameters(businessError.parameters())
-            .build();
-
-    final ErrorResponseList errorResponseList =
-        ImmutableErrorResponseList.builder()
-            .serviceName(serviceName)
-            .addAllFaults(List.of(errorResponse))
+    final ErrorResponseFault errorResponseFault =
+        ImmutableErrorResponseFault.builder()
+            .code(businessError.errorCode().toString())
+            .message(businessError.errorCode().template())
+            .addAllMessageParameters(businessError.parameters())
             .build();
 
     return ResponseEntity.status(businessError.errorCode().errorCodeType().getCode())
-        .body(errorResponseList);
+        .body(
+            ImmutableErrorResponse.builder()
+                .serviceName(serviceName)
+                .addAllErrorResponseFaults(List.of(errorResponseFault))
+                .build());
   }
 }