+++ domain/src/main/java/com/di/hostme/service/domain/country/query/FindCountryByIdQuery.java
@@ -1,0 +1,11 @@
+package com.di.hostme.service.domain.country.query;
+
+import java.util.UUID;
+import org.immutables.value.Value;
+
+@Value.Immutable
+public interface FindCountryByIdQuery {
+
+  @Value.Parameter
+  UUID id();
+}