+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 ## (template)
 ### ⭐ New Features
-  - No changes.
+  - *[KAN-3]* Added project skeleton
+  - *[KAN-9]* Added Apache License
+  - *[KAN-5]* Implemented `GET v1/api/host-me/countries/{countryId}` and `GET v1/api/host-me/countries` endpoints
 ### 🐞 Bug Fixes
   - No changes.
 ### 📔 Documentation